Make the username starts with @ symbol. Fixed FileSize problem.

+++ SuttorLib/Models/Book.cs
@@ -16,7 +16,7 @@ namespace SuttorLibrary.Models
         
         public string FilePath { get; set; }
         public string FileType { get; set; }
-        public long FileSize { get; set; }
+        public double FileSize { get; set; }
         public string PhotoPath { get; set; }
         public DateTime UploadedAt { get; set; }
